Add crude but useful critical move finder
Really, it finds the worst blunder made by the losing player that
arguably lost the game.  The heuristic is simple: the analysis
proceeds backwards from the end of the game, looking for the move
with the highest change in score.  This does mean that a lot of
the middle game is often skipped, and this really just provides
the point at which the game was already won, and all that was
needed was to play winning moves.  However, this is still good
practice for most of us...

For now, it's just a command line program and it spits out some
debugging junk.