Add support for showing the "critical moment"
Currently this is just shown in the game listing.  It is derived
from analysis stored in a JSON file with the same base name as
one of the builtin PGN files.  The JSON is the output from
crit_finder.rb (also enhanced in this commit for better machine-
readability).  In future I'd like to add support for skipping
directly to the critical move.