Fix handling of correctly guessed en-passant
The captured pawn needs to be removed.  Also, make our unit tests
more realistic by using the real Position model class.  This makes
the mock Chessboard class a hybrid, but the main point is to skip
over parts that update the display and assert that the pieces are
on the expected squares.