Builtins: use basename of the filename in the URI
For example, if the PGN file is `anderssen-games.pgn`, the URI is
`/builtin/anderssen`.  The first game in this collection is
`/builtin/anderssen/0`, etc.  The intent is to ensure that if
new PGN files are added to the built-in collection, this does
not affect the URIs of existing built-ins.

In order to avoid opening up any security vulnerability for people
who try to probe the underlying filesystem, we use an allow list
with a map, so actual names of files are not in the request,
nor can any files not explicitly mapped be accessed.